Move parse error into its own file
