+++ parser/src/lib.rs
@@ -1,4 +1,4 @@
-use lalrpop_util::{lalrpop_mod, lexer::Token};
+use lalrpop_util::lalrpop_mod;
 
 pub mod ast;
 pub use ast::*;
@@ -6,128 +6,10 @@ pub use ast::*;
 pub mod parse;
 pub use parse::*;
 
-use std::{fmt::Display, iter};
+pub mod error;
+pub use error::ParseError;
 
 lalrpop_mod!(#[allow(clippy::all)] pub parser, "/cel.rs");
-
-#[derive(Debug, Clone)]
-pub struct Location {
-    pub line: usize,
-    pub column: usize,
-    pub absolute: usize,
-}
-
-#[derive(Debug, Default)]
-pub struct Span {
-    pub start: Option<Location>,
-    pub end: Option<Location>,
-}
-
-impl Span {
-    fn single_location(location: Location) -> Span {
-        Span {
-            start: Some(location.clone()),
-            end: Some(location),
-        }
-    }
-}
-
-#[derive(Debug)]
-pub struct ParseError {
-    pub msg: String,
-    pub expected: Vec<String>,
-    pub span: Span,
-}
-
-impl Display for ParseError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str(&self.msg)
-    }
-}
-
-impl ParseError {
-    fn from_lalrpop(src_str: &str, err: lalrpop_util::ParseError<usize, Token<'_>, &str>) -> Self {
-        use lalrpop_util::ParseError::*;
-
-        match err {
-            InvalidToken { location } => ParseError {
-                span: byte_pos_to_src_location(src_str, location)
-                    .map(Span::single_location)
-                    .unwrap_or_default(),
-                expected: Vec::new(),
-                msg: "invalid token".into(),
-            },
-            UnrecognizedEof { location, expected } => ParseError {
-                msg: "unrecognized eof".into(),
-                span: byte_pos_to_src_location(src_str, location)
-                    .map(Span::single_location)
-                    .unwrap_or_default(),
-                expected,
-            },
-            UnrecognizedToken {
-                token: (start, token, end),
-                expected,
-            } => ParseError {
-                msg: format!("unrecognized token: '{}'", token),
-                span: Span {
-                    start: byte_pos_to_src_location(src_str, start),
-                    end: byte_pos_to_src_location(src_str, end),
-                },
-                expected,
-            },
-            ExtraToken {
-                token: (start, token, end),
-            } => ParseError {
-                msg: format!("extra token: '{}'", token),
-                span: Span {
-                    start: byte_pos_to_src_location(src_str, start),
-                    end: byte_pos_to_src_location(src_str, end),
-                },
-                expected: Vec::new(),
-            },
-            User { error } => ParseError {
-                msg: error.into(),
-                expected: Vec::new(),
-                span: Span::default(),
-            },
-        }
-    }
-}
-
-// Slightly simplified but heavily inspired by
-// https://github.com/gluon-lang/gluon/blob/f8326d21a14b5f21d203e9c43fa5bb7f0688a74c/base/src/source.rs
-fn byte_pos_to_src_location(src_str: &str, byte_pos: usize) -> Option<Location> {
-    let src_bytes = src_str.as_bytes();
-    let total_len = src_bytes.len();
-
-    let line_indices: Vec<usize> = {
-        let input_indices = src_bytes
-            .iter()
-            .enumerate()
-            .filter(|&(_, b)| *b == b'\n')
-            .map(|(i, _)| i + 1); // index of first char in the line
-
-        iter::once(0).chain(input_indices).collect()
-    };
-
-    if byte_pos <= total_len {
-        let num_lines = line_indices.len();
-
-        let line_index = (0..num_lines)
-            .find(|&i| line_indices[i] > byte_pos)
-            .map(|i| i - 1)
-            .unwrap_or(num_lines - 1);
-
-        let line_byte_pos = line_indices[line_index];
-        Some(Location {
-            line: line_index,
-            column: byte_pos - line_byte_pos,
-            absolute: byte_pos,
-        })
-    } else {
-        None
-    }
-}
 
 /// Parses a CEL expression and returns it.
 ///
